arm/mach-exynos(4): Rename mach-exynos4 back to mach-exynos
mach-exynos in the Samsung GT-I9100 source tree drop contained a tree
significantly ahead of the mainline kernel, and even farther ahead
of the 3.0.x Linux series, which has not seen an update to mach-exynos(4)
since July 2011.

The mach-exynos tree was improperly renamed from the source drop prior to
committing to git, causing the original rename to not show in git history.
The rename was done in an attempt to improve mainline commonality, however:
1)  Even with the rename, any mainline patches that touch mach-exynos(4) have
no chance of cleanly applying to the Exynos tree due to significant differences
in the directory contents
2)  It just doesn't matter - see above, this tree has not seen a mainline commit
since August 2011, and it is now March 2012.  The rename fragments this tree
from ones that actually will have active development on mach-exynos, compared
to mainline which has had none since 2011 on 3.0.x
3)  It causes any patches coming from other I9100 community development to
fail to apply to this source tree.

For the above reasons, the benefits of mach-exynos outweigh those of mach-exynos4
and similarly, the disadvantages of mach-exynos4 outweigh those of mach-exynos,
and the tree should be renamed.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>