exynos4: Disable SW downclock in cpuidle
The Samsung 3.0.x source includes a new exynos4 feature:
reducing CPU clock when in idle.  However, this seems
pointless as the CPU clock gets gated in this state anyway.

Also, the logic for doing this seems to sometimes hang,
as there are occasional cases of the RCU stall detector
going off due to CPU1 getting stuck in exynos4_enter_idle()

This issue seems to go away when disabling the SW downclock
feature with not much negative effect on battery, so disable
it for now until it can be revisited.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>